+++ lib/src/widgets/getx/playx_material_app.dart
@@ -1,15 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
-import 'package:playx/src/widgets/models/playx_app_settings.dart';
-import 'package:playx/src/widgets/models/playx_screen_settings.dart';
-import 'package:playx/src/widgets/models/playx_theme_settings.dart';
+import 'package:playx/src/models/playx_app_settings.dart';
+import 'package:playx/src/models/playx_screen_settings.dart';
+import 'package:playx/src/models/playx_theme_settings.dart';
 import 'package:playx_localization/playx_localization.dart';
 import 'package:playx_theme/playx_theme.dart';
 import 'package:playx_widget/playx_widget.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
-import '../models/playx_get_navigation_settings.dart';
+import '../../models/playx_get_navigation_settings.dart';
+
+
 
 /// [PlayxGetMaterialApp] is a widget that integrates [GetMaterialApp] with [PlayXThemeBuilder] and [ScreenUtilInit].
 ///